proper header support for client and server
