+++ src/base.js
@@ -2,6 +2,7 @@ var EventEmitter = require('events').EventEmitter;
 var NamespaceContext = require('./parser/nscontext');
 var SOAPElement = require('./soapModel').SOAPElement;
 var xmlBuilder = require('xmlbuilder');
+var XMLHandler = require('./parser/xmlHandler');
   
 class Base extends EventEmitter {
   constructor(wsdl, options) {
@@ -120,6 +121,28 @@ class Base extends EventEmitter {
       nsContext.addNamespace(prefix, nsURI);
     }
     return nsContext;
+  }
+
+  addSoapHeadersToEnvelope(soapHeaderElement, xmlHandler) {
+    for (let i = 0, n = this.soapHeaders.length; i < n; i++) {
+      let soapHeader = this.soapHeaders[i];
+      let elementDescriptor;
+      if (typeof soapHeader.value === 'object') {
+        if (soapHeader.qname !== undefined) {
+          if (soapHeader.qname.nsURI !== undefined) {
+            let element = this.findElement(soapHeader.qname.nsURI, soapHeader.qname.name);
+            elementDescriptor =
+              element && element.describe(this.wsdl.definitions);
+          }
+        }
+        xmlHandler.jsonToXml(soapHeaderElement, null, elementDescriptor,
+          soapHeader.value);
+      } else { //soapHeader has XML value
+        XMLHandler.parseXml(soapHeaderElement, soapHeader.xml);
+        xmlHandler.jsonToXml(soapHeaderElement, null, null,
+          soapHeader.value);
+      }
+    }
   }
 
 }
